+++ src/main/java/ai/fullfillment_v2/inbound/domain/InboundProduct.java
@@ -1,4 +1,4 @@
-package ai.fullfillment_v2.inbound;
+package ai.fullfillment_v2.inbound.domain;
 
 import org.springframework.util.Assert;
