+++ src/main/java/ai/fullfillment_v2/inbound/domain/InboundStatus.java
@@ -1,4 +1,4 @@
-package ai.fullfillment_v2.inbound;
+package ai.fullfillment_v2.inbound.domain;
 
 public enum InboundStatus {
     ORDER_REQUESTED("발주 요청");
