+++ src/main/java/ai/fullfillment_v2/inbound/feature/CreateInbound.java
@@ -1,5 +1,7 @@
-package ai.fullfillment_v2.inbound;
+package ai.fullfillment_v2.inbound.feature;
 
+import ai.fullfillment_v2.inbound.domain.Inbound;
+import ai.fullfillment_v2.inbound.domain.InboundProduct;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
